+++ Vault/src/utopia/vault/nosql/targeting/many/TargetingMany.scala
@@ -2,6 +2,7 @@ package utopia.vault.nosql.targeting.many
 
 import utopia.flow.generic.model.immutable.Value
 import utopia.vault.nosql.targeting.Targeting
+import utopia.vault.nosql.targeting.one.TargetingOne
 
 object TargetingMany
 {
@@ -13,12 +14,15 @@ object TargetingMany
 	// NESTED   -------------------------
 	
 	private class Wrapper[O, +R](override val wrapped: TargetingMany[O])(f: O => R)
-		extends TargetingMany[R] with TargetingManyWrapper[TargetingMany[O], O, R, TargetingMany[R]]
+		extends TargetingMany[R]
+			with TargetingManyWrapper[TargetingMany[O], TargetingOne[Option[O]], O, R, TargetingMany[R], TargetingOne[Option[R]]]
 	{
 		override protected def self: TargetingMany[R] = this
 		override protected def mapResult(result: O): R = f(result)
 		
 		override protected def wrap(newTarget: TargetingMany[O]): TargetingMany[R] = new Wrapper(newTarget)(f)
+		override protected def wrapUniqueTarget(target: TargetingOne[Option[O]]): TargetingOne[Option[R]] =
+			target.mapResult { _.map(f) }
 	}
 }
 
@@ -27,7 +31,8 @@ object TargetingMany
   * @author Mikko Hilpinen
   * @since 18.05.2025, v1.21
   */
-trait TargetingMany[+A] extends Targeting[Seq[A], Seq[Value]] with TargetingManyLike[A, TargetingMany[A]]
+trait TargetingMany[+A]
+	extends Targeting[Seq[A], Seq[Value]] with TargetingManyLike[A, TargetingMany[A], TargetingOne[Option[A]]]
 {
 	/**
 	  * @param f A mapping function applied to this access point's results