+++ Vault/src/utopia/vault/nosql/targeting/one/TargetingOne.scala
@@ -1,12 +1,15 @@
 package utopia.vault.nosql.targeting.one
 
 import utopia.flow.generic.model.immutable.Value
-import utopia.vault.nosql.targeting.Targeting
+import utopia.vault.nosql.targeting.many.TargetingManyLike
+import utopia.vault.nosql.targeting.{Targeting, TargetingWrapper}
 
 // TODO: Possibly, add implicitly map function for instances which yield an option
 object TargetingOne
 {
 	// OTHER    ----------------------------
+	
+	def headOf[T <: TargetingManyLike[A, T, _], A](target: T): TargetingOne[Option[A]] = new Head[T, A](target)
 	
 	def map[O, R](original: TargetingOne[O])(f: O => R): TargetingOne[R] = new Wrapper[O, R](original, f)
 	
@@ -20,6 +23,17 @@ object TargetingOne
 		
 		override protected def wrapResult(result: O): R = f(result)
 		override protected def wrap(newTarget: TargetingOne[O]): TargetingOne[R] = new Wrapper(newTarget, f)
+	}
+	
+	private class Head[T <: TargetingManyLike[A, T, _], A](override val wrapped: T)
+		extends TargetingOne[Option[A]] with TargetingWrapper[T, Seq[A], Seq[Value], Option[A], Value, TargetingOne[Option[A]]]
+	{
+		override protected def self: TargetingOne[Option[A]] = this
+		
+		override protected def wrapResult(result: Seq[A]): Option[A] = result.headOption
+		override protected def wrapValue(value: Seq[Value]): Value = value.headOption.getOrElse(Value.empty)
+		
+		override protected def wrap(newTarget: T): TargetingOne[Option[A]] = new Head[T, A](newTarget)
 	}
 }